This update brings a bunch of iPad support that we've developed for our applications, as well as various extensions and fixes for the Mac. The existing iPhone/iPad code has been wrapped up into static library targets for easier use (of course, only a subset of the code has been ported to iPhone OS, as needed).
Details follow:

OmniUI/iPad:
  - First release of this framework!
  - Includes a bunch of features from our iPad apps
    - Document picker, basic document-based application support
    - CoreText-based text editor (still a work in progress, but a great start)
      - We'll be working more on performance for large blobs of text, which we haven't needed yet.
    - Zooming/tiling scroll view system
    - Extensible inspector system; includes full font and color inspectors (palette, hsv, rgb, gray)
    - widgets (thumb wheel, color swatches, toggle buttons, matching segmented control, editable text well)
  - Support for sending feedback, exporting documents/PDF via Mail, documentation, etc.
  - See examples in OmniGroup/Frameworks/OmniUI/iPad/Examples
    - The text editor example is still in progress, but isn't fully hooked up yet (no inspectors yet)
  - Gesture recognizer debugging helper (logs all events and state changes)
  - "Direct" tap recognizer subclass
  - Overlay hint view (for example, showing the % zoom level while in the middle of a zoom gesture).
  - Basic RTF reading/writing

OmniQuartz:
  - First release of this framework!
  - Includes a subset as a static library OmniQuartzTouch target for iPhone OS.
  - Includes lots of CALayer debugging support, helpers (log all animations, logging of layer trees, vector output for layer trees)
  - Includes some basic drawing routings for
  - Includes some simple CoreImage utilities
  - Includes a class cluster of color component preserving color classes. UIColor is more "build-only". OQColor remembers the components you gave it and acts more like a model object (including XML and plist archiving support based on NSColor(OAExtensions)).
  - Includes delayed layer removal animations (since removing a layer normally is instantaneous in CA unless you want default animation).
  - Includes a (perspective correct) view transition flip animation
  - Includes a 'hole' layer.

Configurations:
  - Consolidated some Mac and iPhone settings such that the Omni-Global-* files are shared between platforms.
  - Added Target-{Mac,Touch}-* configurations for target-specific options
  - Added support for static libraries on the iPhone, that act somewhat like frameworks

General:
  - Various clang + llvm 2.7 static warnings and static analyzer fixes
  - Made various bits of constant data be marked 'const' (notification names, dictionary keys, etc)
  - More work on GC support. Still not there.
  - Switched from NS geometry types to CG types in many places. -DNS_BUILD_32_LIKE_64 is your friend.
  - Rearranged/refactored extensions into separate files to allow porting smaller relevant chunks to iPhone OS

Scripts:
  - Added a script to export a strings file based on the files that were compiled into a static library. It's a hack.
  - Added a script that can be run from an iPad app to merge resources exported by the static library projects. See the "Copy Library Resources" script phase in OmniGroup/Frameworks/OmniUI/iPad/Examples/TextEditor for an example.

FixStringsFile:
  - Added this little tool we use to sort and do basic ASCII->Unicode substitution on strings files ("..." becomes an ellipsis, for example).

OmniBase:
  - Added OmniBaseTouch static library target for iPhone/iPad
  - Removed OBError/OBErrorWithInfo macros. These weren't suitable for iPad development where OMNI_BUNDLE_IDENTIFIER might be the main bundle. Each framework domain should define its own macros and own domain global NSString key.
  - Added OBLoadBundle tool. This is is useful to run the OmniBase ABI consistency checks on arbitrary frameworks/bundles.
  - Added OBAssertThrows, which is just like STAssertThrows but doesn't provoke a clang-sa warning.
  - Added scanned memory allocation wrapper macros that do a reasonable job of working when GC is enabled or not.
  - Make OMNI_BUNDLE work on iPad (just trivially the main bundle in that case).
  - Added OBFinishPortingLater(). This will warn once when hit.
  - Added _OBIsNegative helpers and OBASSERT_NONNEGATIVE macro. This is useful for avoiding clang warnings about comparing unsigned vs. zero.
  - Define the NS_RETURNS_RETAINED clang-sa helper macro, if it isn't defined already since iPhone OS 3.2 doesn't define this.
  - Turn off RCS_ID strings on iPhone. No one is going to run ident on the executable and the executable size matters more here.

OmniFoundation:
  - Added OmniFoundationTouch with the common set of classes/extensions our apps need. You might want to tweak it for your own use.
  - Disable libbz2 compression support on iPhone OS. libbz2 is present, but the headers aren't and Apple considers this private API.
  - Split out a OFEnumNameTable(OFXMLArchiving) category in separate files. Nice if you don't want any of the XML support in your library.
	- Added -[OFMultiValueDictionary parametersFromQueryString] for parsing URL query strings.
	- OFFilterProcess uses POSIX spawn() if possible. Also fixed some GC issues.
  - Added OFNetChangeNotifier.
  - Added -[OFPreference coerceStringValue:toTypeOfPropertyListValue] -- used in the preference setting URL support in our iPhone OS apps
  - Added OFSaveType; a mirror of NSSaveOperation, but available on iPad.
  - Added more block-based helpers on NSArray, -indexByBlock, -firstInRange:that:, -last:, -lastInRange:that:
  - Added ICS-formatted NSDate/NSString conversion methods in NSDate(OFExtensions).
  - Added block-based -[NSString(OFExtensions) strinByPerformingReplacement:onCharacters:options:range:].
  - Added OFASN1DescribeOID()

OmniFoundation/OmniAppKit:
  - Added some stand-in classes or stubs for AppKit classes. More refinement to come.

	  OFFileWrapper
    OAFontDescriptor
    OAParagraphStyle
    OAMutableParagraphStyle
    OATextTab

  - New OmniAppKitTouch target with just these stand-ins.
  - New OAStepperTextField control.
  - Removed deprecated support for document backups using resource forks.
  - Removed support for fixing up HFS attributes on written file wrappers. This bug has been fixed in the system frameworks for a while.
  - Added -[NSView(OAExtensions) enclosingViewOfClass:]
  - Various tweaks to rendering in OACalendarView

OmniUnzip:
  - Added an OmniUnzipTouch target

OmniSoftwareUpdate:
  - Support for opting in/out of update tracks.
  - Various UI fixes and cleanup
  - More work on signed feeds
  - Set quarantine on installed apps (until we verify their signature)
  - Fixes on the install path
  - Added utilities for generating feeds

+++ Frameworks/OWF/Content.subproj/OWDataStreamCharacterCursor.m
@@ -252,7 +252,7 @@ static /* inline */ NSString *OWCreateStringFromData(NSData *data, CFStringEncod
             case kTextMalformedInputErr:
             case kTECUnmappableElementErr:
 #ifdef DEBUG
-                NSLog(@"Skipping byte at offset %d, value %d ('%c'): ConvertFromTextToUnicode() returned %s (encoding=%@)", input - sourceBytes, (unsigned int)*input, *input, NameForTECStatus(status), CFStringConvertEncodingToIANACharSetName(stringEncoding));
+                NSLog(@"Skipping byte at offset %ld, value %d ('%c'): ConvertFromTextToUnicode() returned %s (encoding=%@)", (long)(input - sourceBytes), (unsigned int)*input, *input, NameForTECStatus(status), CFStringConvertEncodingToIANACharSetName(stringEncoding));
 #endif
                 input++;
 #ifdef MARK_MALFORMED_INPUT
