+++ Frameworks/OmniAppKit/Widgets.subproj/OASteppableTextField.h
@@ -1,4 +1,4 @@
-// Copyright 2006-2007 Omni Development, Inc.  All rights reserved.
+// Copyright 2006-2007,2010 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -11,7 +11,7 @@
 
 @interface OASteppableTextField : NSTextField 
 {
-    int stepperTracking;
+    double stepperTracking;
 }
 
 - (void)stepperAction:(id)sender;
